+++ internal/errors/errors.go
@@ -24,4 +24,5 @@ var (
 	ErrNoPortsFound        = errors.New("no neutron ports found")
 
 	ErrSnatPoolSizeUnsupportedProvider = errors.New("snat_pool_size is only supported for provider=f5")
+	ErrSnatIPConflict                  = errors.New("SNAT IP conflict")
 )